requirements.txt
chromadb>=0.5.0
google-generativeai>=0.5.0
sentence-transformers>=2.7.0
pypdf>=4.0.0
requests>=2.31.0
pytest>=8.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
